% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUp.R
\name{cleanUp}
\alias{cleanUp}
\title{\code{cleanUp} cleans up TrackSig folder for the next round of usage
\code{cleanUp} store results and generated plot (with preferred name, plot_name) 
to the given directory, result_path, delete results and saved_data in TrackSig folder.}
\usage{
cleanUp(result_path, plot_name)
}
\arguments{
\item{result_path}{the absolute path of where the pdf plot and results will be stored}

\item{plot_name}{preffered plot name by user}
}
\description{
\code{cleanUp} cleans up TrackSig folder for the next round of usage
\code{cleanUp} store results and generated plot (with preferred name, plot_name) 
to the given directory, result_path, delete results and saved_data in TrackSig folder.
}
\examples{
result_path <- "c://User//cindy//Documents//TrackSig_results"
cleanUp(result_path, "tumor_31.pdf")
}
