% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mutational_signatures_multiple.R
\name{compute_signatures_for_all_examples}
\alias{compute_signatures_for_all_examples}
\title{\code{compute_signatures_for_all_examples} Generate activitracjectories of mutational signatures with patients
with multiple samples
\code{compute_signatures_for_all_examples} Please see:
\url{https://github.com/YuliaRubanova/TrackSig/blob/master/src/compute_mutational_signatures.R}}
\usage{
compute_signatures_for_all_examples(dir_counts = DIR_COUNTS)
}
\arguments{
\item{dir_counts}{default data/counts generated by make_counts.sh}
}
\description{
\code{compute_signatures_for_all_examples} Generate activitracjectories of mutational signatures with patients
with multiple samples
\code{compute_signatures_for_all_examples} Please see:
\url{https://github.com/YuliaRubanova/TrackSig/blob/master/src/compute_mutational_signatures.R}
}
\examples{
compute_signatures_for_all_examples()
}
