% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUp.R
\name{downloadhg19}
\alias{downloadhg19}
\title{\code{downloadhg19} Download hg 19 reference from the internet
\code{downloadhg19} downloads hg 19 reference from the internet 
and unzip the fa.gz files to make it ready to use for TrackSig. 
For the usage of TrackSig, it is required to place the hg19 reference 
in a folder called hg19 in annotation directory of TrackSig. This will 
be done by this function.\strong{bold}: This function will take quite 
a bit of time. This function is inspired by Rich Scriven on 
\url{https://stackoverflow.com/questions/33790052/download-all-files-from-a-folder-on-a-website}}
\usage{
downloadhg19()
}
\description{
\code{downloadhg19} Download hg 19 reference from the internet
\code{downloadhg19} downloads hg 19 reference from the internet 
and unzip the fa.gz files to make it ready to use for TrackSig. 
For the usage of TrackSig, it is required to place the hg19 reference 
in a folder called hg19 in annotation directory of TrackSig. This will 
be done by this function.\strong{bold}: This function will take quite 
a bit of time. This function is inspired by Rich Scriven on 
\url{https://stackoverflow.com/questions/33790052/download-all-files-from-a-folder-on-a-website}
}
\examples{
downloadhg19()
}
\seealso{
The hg 19 reference can also be downloaded here:
\url{http://hgdownload.cse.ucsc.edu/goldenPath/hg19/chromosomes/}
}
