% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUp.R
\name{generatePlots}
\alias{generatePlots}
\title{\code{generatePlots} gathers all the function calls and generates plot.
\code{generatePlots} generates one pdf plot for patient with multiple sample
so that it is easier for intepretation.}
\usage{
generatePlots()
}
\description{
\code{generatePlots} gathers all the function calls and generates plot.
\code{generatePlots} generates one pdf plot for patient with multiple sample
so that it is easier for intepretation.
}
\examples{
generatePlots()
}
