% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlotTitle.R
\name{getPlotTitle}
\alias{getPlotTitle}
\title{\code{getPlotTitle} output a sequence of logarithmically spaced numbers.}
\usage{
getPlotTitle(plot_name)
}
\arguments{
\item{plotname}{the path of where the pdf plot result will be stored 
which is generated by compute_signatures_for_all_examples funtion in 
plotting.R in TrackSig/scr}

\item{to}{last number in the sequence.}

\item{length.out}{length of the sequence, a positive integer > 1. Default is 10.}
}
\value{
A string of the tilte of the plots.
}
\description{
\code{getPlotTitle} generates plot title for samples
}
\examples{
plot_name <- "results_signature_trajectories/ALL/sampleId/sampleId.pdf"
getPlotTitle(plot_name)
}
\seealso{
\code{\link{compute_signatures_for_all_examples}}
}
