% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUp.R
\name{importTrackSig}
\alias{importTrackSig}
\title{\code{importTrackSig} fork TrackSig framework onto local repository from Git Hub.}
\usage{
importTrackSig()
}
\description{
\code{importTrackSig} fork TrackSig framework onto local repository from Git Hub.
}
\examples{
importTrackSig()
}
\seealso{
\url{https://github.com/YuliaRubanova/TrackSig.git}
}
