% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUp.R
\name{makeMutationCounts}
\alias{makeMutationCounts}
\title{\code{makeMutationCounts}, a commandline from TrackSig framework to generates information 
about order of mutation occurence, mutation types, mutation counts and bootstrapping result
for signaficance extimation.
\code{makeMutationCounts} places all the generated infomation under "TrackSig/data". 
Those information will be used to generate plots.}
\usage{
makeMutationCounts()
}
\description{
\code{makeMutationCounts}, a commandline from TrackSig framework to generates information 
about order of mutation occurence, mutation types, mutation counts and bootstrapping result
for signaficance extimation.
\code{makeMutationCounts} places all the generated infomation under "TrackSig/data". 
Those information will be used to generate plots.
}
\examples{
makeMutationCounts()
}
