% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUp.R
\name{makeVaf}
\alias{makeVaf}
\title{\code{makeVaf} Use command line to convert a vcf file of point mutation to a vaf file of variant 
allel frequency.}
\usage{
makeVaf()
}
\description{
\code{makeVaf} Use command line to convert a vcf file of point mutation to a vaf file of variant 
allel frequency.
}
\examples{
makeVaf()
}
\seealso{
\url{https://github.com/YuliaRubanova/TrackSig}
}
