% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutSig.R
\name{MutSig}
\alias{MutSig}
\title{\code{MutSig} plot mutational signature activity of given simple(s)}
\usage{
MutSig(sigActPoints, changepoints, phis, sigActPoints2 = NULL,
  changepoints2 = NULL, phis2 = NULL)
}
\arguments{
\item{sigActPoints}{absolute path of a CSV file of signature activities 
across pesudotime for each signature}

\item{changepoints}{absolute path of a txt file containing the significant 
changepoints of signature activities}

\item{phis}{absolute path of txt containing the pesudo-timeline}

\item{sigActPoints2}{an optional file for two seamples plot. See 
sigActPoints for the formate}

\item{changepoints2}{an optional file for two seamples plot. See 
changepoint for the formate}

\item{phis2}{an optional file for two seamples plot. See phis for the 
formate}
}
\value{
a list containg a pesudotime (estimated by number of mutants per 
cancer cell) vs.signature activity plot and signatures. For two chronological 
samples, it returns list containg above information for each sample
}
\description{
\code{MutSig} takes three input file (a .csv file of signature activities 
corrdinate where x is the esudotime and y is the signature activity, 
a .txt file of signaficant signature activity change point, and a .txt 
file for the pesudo-timeline) and generates a plot of a pesudotime 
(extimated by mutant allel per cell) vs. signature activities for single 
sample. For two two chronological samples, it takes six input file 
and generate the plot. This function is inspired by TrackSig. 
Please see reference at the top of this file
}
\examples{
# For single sample
sigActPoints <- system.file("extdata", "sigActPoints.csv", package = "SigAct")
changepoints <- system.file("extdata", "changepoints.txt", package = "SigAct")
phis <- system.file("extdata", "phis.txt", package = "SigAct")
MutSig(sigActPoints, changepoints, phis)
# For two samples
sigActPoints2 <- system.file("extdata", "sigActPoints2.csv", package = "SigAct")
changepoints2 <- system.file("extdata", "changepoints2.txt", package = "SigAct")
phis2 <- system.file("extdata", "phis2.txt", package = "SigAct")
MutSig(sigActPoints, changepoints, phis, sigActPoints2, changepoints2, phis2)
}
