% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutSig.R
\name{plotSignatures}
\alias{plotSignatures}
\title{\code{plotSignatures} Plot signatre activity plot}
\usage{
plotSignatures(sigActPoints, phis = NULL, changepoints = NULL,
  ytitle = "Signature exposure (\%)",
  xtitle = "Avg number of mutant alleles per cancer cell",
  removeSigsBelow = 0, sig_colors = NULL)
}
\arguments{
\item{sigActPoints}{data frame of signature activities across pesudotime for all signatures 
in the sample}

\item{phis}{character verctor representing the pesudo-timeline}

\item{changepoints}{character vector for significant changepoints of signature 
activities}

\item{ytitle}{title of y-axis}

\item{xtitle}{title for x-axis}

\item{removeSigsBelow}{threshold of signature activities. Only signature activity above this
threshold will be plotted}

\item{sig_colors}{customized colour set for signatures}
}
\value{
A list containg a plot with pesudo-timeline as the x-axis and 
signative activities (in percentage) as the y-axis and the mutational 
signatures
}
\description{
\code{plotSignatures} This function take processed information and generate 
a plot with pesudo-timeline as the x-axis and signative activities (in 
percentage) as the y-axis. Helper Function for MutSig. This function is 
modified from plot_signatures function of plotting.R TrackSig 
Please see reference at the top of this file
}
