% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutSig.R
\name{readFiles}
\alias{readFiles}
\title{\code{readFiles} Read and Process MutSig files for plotting}
\usage{
readFiles(sigActPoints, changepoints, phis)
}
\arguments{
\item{sigActPoints}{absolute path of a CSV file of signature activities 
across pesudotime for each signature}

\item{changepoints}{absolute path of a txt file containing the significant 
changepoints of signature activities}

\item{phis}{absolute path of txt containing the pesudo-timeline}
}
\value{
A list constining data frame for signature activity points for plotting, 
a character vector with timepoint where signature activities changed significantly, 
and a vector of character indicated the number of mutant in each cancer cell for 
this simple, which will be use as the pesudo-timeline
}
\description{
\code{readFiles} This helper function for MutSig reads and processes files 
for plotting. Helper Function for MutSig.This function is based on 
the compute_signatures_for_all_examples function in compute_mutational_signatures.R 
and the read_mistures function from helper_functions.R. 
Please see reference at the top of this file
}
