% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigProposedEtiology.R
\name{readSignatures}
\alias{readSignatures}
\title{\code{readSignatures} provide information about signatures' proposed etiology}
\usage{
readSignatures()
}
\value{
a data frame of proposed etiology of mutational signature with conlums: signatures, 
Proposed_etiology and detailed information of signature etiology
}
\description{
\code{readSignatures} takes external file of signatures' proposed etiology and 
convert it to R data frame. Helper function for sigProposedEtiology.
}
