% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMultiApp.R
\name{runMultiApp}
\alias{runMultiApp}
\title{\code{runRptApp} launch the shiny app distributed with this package framework}
\usage{
runMultiApp()
}
\description{
\code{runRptApp} launches the shiny app for which the code has been placed in  \code{./inst/shiny-scripts/rptApp/}.
}
