% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mutational_signatures_multiple.R
\name{save_data_for_samples}
\alias{save_data_for_samples}
\title{\code{save_data_for_samples} Prepare all the data needed for plots}
\usage{
save_data_for_samples(dir_counts = DIR_COUNTS,
  bootstrap_counts = BOOTSTRAP_COUNTS)
}
\arguments{
\item{dir_counts}{default data/counts generated by make_counts.sh}

\item{bootstrap_counts}{default data/bootstrap generated by make_counts.sh}
}
\description{
\code{save_data_for_samples} Please see 
\url{https://github.com/YuliaRubanova/TrackSig/blob/master/src/compute_mutational_signatures.R}
}
\examples{
save_data_for_samples()
}
