% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mutational_signatures_multiple.R
\name{setUpCompute}
\alias{setUpCompute}
\title{This is modified based on TrackSig's oringinal compute_mutational_signatures.R
\code{setUpCompute} source the TrackSig/src/header.R code as required for TrackSig}
\usage{
setUpCompute(path)
}
\arguments{
\item{path}{absolute path of header.R in your TrackSig/src folder}
}
\description{
\code{setUpCompute} The framework of TrackSig requires to header.R to be sourced 
to run the following functions.
}
\examples{
path <- "c://User//cindy//Documents//BCB410//TrackSig//src//header.R"
setUpCompute(path)
}
