% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigProposedEtiology.R
\name{sigProposedEtiology}
\alias{sigProposedEtiology}
\title{\code{setUpCompute} source the TrackSig/src/header.R code as required for TrackSig}
\usage{
sigProposedEtiology(sigName, mutationalSig)
}
\arguments{
\item{path}{absolute path of header.R in your TrackSig/src folder}
}
\description{
\code{setUpCompute} The framework of TrackSig requires to header.R to be sourced
to run the following functions.
}
\examples{
path <- "c://User//cindy//Documents//BCB410//TrackSig//src//header.R"
setUpCompute(path)
}
