% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mutational_signatures_multiple.R
\name{generatePlots}
\alias{generatePlots}
\title{\code{generatePlots} Generate activitracjectories of mutational signatures with patients
with multiple samples
\code{generatePlots} Please see:
\url{https://github.com/YuliaRubanova/TrackSig/blob/master/src/compute_mutational_signatures.R}}
\usage{
generatePlots()
}
\description{
\code{generatePlots} Generate activitracjectories of mutational signatures with patients
with multiple samples
\code{generatePlots} Please see:
\url{https://github.com/YuliaRubanova/TrackSig/blob/master/src/compute_mutational_signatures.R}
}
\examples{
generatePlots()
}
