% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigProposedEtiology.R
\name{readSignatures}
\alias{readSignatures}
\title{\code{readSignatures} provide information about signatures}
\usage{
readSignatures(path = "./mutational_signatures.txt")
}
\arguments{
\item{path}{absolute path external file of mutational signature information}
}
\value{
a data frame of mutational signatures with conlums: signatures, 
Proposed_etiology and detailed information of signature etiology
}
\description{
\code{readSignatures} takes external file of signature information and 
convert it to R data frame
}
\examples{
path <- "./mutational_signatures.txt"
mutationalSig <- readSignatures(path)
}
