% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigProposedEtiology.R
\name{sigProposedEtiology}
\alias{sigProposedEtiology}
\title{\code{sigProposedEtiology} Give proposed etiology of mutational signature}
\usage{
sigProposedEtiology(sigName)
}
\arguments{
\item{sigName}{ID of a signature}
}
\description{
\code{sigProposedEtiology} This function takes signature ID and give the 
proposed etiology of mutational signature
}
\examples{
sigProposedEtiology("S1")
}
